requirements.txt
torch
torchvision
pandas
matplotlib
scikit-learn
tqdm
opencv-python
Pillow